refactor(rag): RAG ライブラリをワークスペースパッケージ化
- RAG 実装と関連スクリプトを packages/rag に移設してパッケージとして分離
- desktop アプリの参照先を @repo/rag に切り替え、workspace 依存を追加
- packages/rag に package.json、tsconfig、oxlint 設定、gitignore を追加
