+++ packages/rag/scripts/example.txt
