+++ packages/rag/scripts/run-retrieve-json.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs/promises'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 
-import { retrieveRagContext } from '../../electron/shared/lib/rag/retrieve.ts'
+import { retrieveRagContext } from '../src/retrieve.ts'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
@@ -14,7 +14,7 @@ main().catch((err) => {
 async function main() {
   const question = '主要機能'
   const result = await retrieveRagContext(question, {
-    dbPath: path.join(__dirname, '..', '..', 'data', 'example.json'),
+    dbPath: path.join(__dirname, '..', 'data', 'example.json'),
     docName: 'example-doc',
     loadChunks: async (dbPath, docName) => {
       const data = await fs.readFile(dbPath, 'utf-8')
